Split off `manager` module parts
